fastapi/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

torch==2.9.1+cu130

transformers==4.57.3
accelerate==1.12.0
bitsandbytes==0.48.2
sentence-transformers==5.1.2

fastapi[standard]==0.123.10
uvicorn[standard]==0.38.0

python-dotenv==1.2.1
